More modules from refactoring self-use.
